backend/back is end/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==7.1.2
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.18.0
Flask==1.1.4
flask-cors==5.0.1
flatbuffers==25.2.10
fpdf==1.7.2
fsspec==2025.3.2
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.31.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
keras==2.15.0
Keras-Preprocessing==1.1.2
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.24.0
libclang==18.1.1
lxml==5.3.2
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.2.1
nltk==3.9.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.1.0
preshed==3.0.9
protobuf==4.25.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-docx==1.1.0
python-Levenshtein==0.24.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.0.2
scipy==1.13.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2+cu118
torchaudio==2.6.0
torchvision==0.17.2+cu118
tqdm==4.67.1
transformers==4.51.3
triton==2.2.0
typer==0.4.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
wasabi==0.10.1
weasel==0.3.4
Werkzeug==1.0.1
wrapt==1.14.1
zipp==3.21.0
